backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

email-validator==2.2.0

SQLAlchemy==2.0.34
alembic==1.13.2
psycopg[binary]==3.2.1

pydantic[email]==2.8.2
pydantic-settings==2.4.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.9

qrcode==7.4.2
Pillow==10.4.0

bcrypt==4.0.1